added authorization

+++ twitter_clone/app/Http/Controllers/IdeaController.php
@@ -4,14 +4,18 @@ namespace App\Http\Controllers;
 
 use App\Models\Idea;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class IdeaController extends Controller
 {
     public function store(Request $request)
     {
+
         $validatedData = $request->validate([
             'content' => 'required|string|max:250',
         ]);
+
+        $validatedData['user_id'] = Auth::user()->id;
 
         Idea::create($validatedData);
         return redirect()->to('/')->with(['success' => 'Idea created successfully']);
@@ -26,6 +30,10 @@ class IdeaController extends Controller
 
     public function edit(Idea $idea)
     {
+        if ($idea->user->id !== Auth::user()->id) {
+            abort(401);
+        }
+
         return view('ideas.edit', [
             'idea' => $idea
         ]);
@@ -33,6 +41,9 @@ class IdeaController extends Controller
 
     public function update(Request $request, Idea $idea)
     {
+        if ($idea->user->id !== Auth::user()->id) {
+            abort(401);
+        }
         $validatedData = $request->validate([
             'content' => 'required|max:250|string'
         ]);
@@ -48,6 +59,9 @@ class IdeaController extends Controller
 
     public function destroy(Idea $idea)
     {
+        if ($idea->user->id !== Auth::user()->id) {
+            abort(401);
+        }
         $idea->delete();
         return redirect()->to('/')->with(['success' => 'Idea deleted successfully']);
     }
